refactoring to use user logic as a module
- created user module
- removed all user logic from auth module to use the new one.
- updated user imports on favorite modules to use the created one
- updated roles numeration imports to point to the user module on game
 module
- updated user imports on reports module to use the created one
