+++ src/app.module.ts
@@ -5,6 +5,7 @@ import { GameModule } from './game/game.module';
 import { AuthModule } from './auth/auth.module';
 import { FavoriteModule } from './favorite/favorite.module';
 import { ReportModule } from './report/report.module';
+import { UserModule } from './user/user.module';
 
 @Module({
   imports: [
@@ -18,6 +19,7 @@ import { ReportModule } from './report/report.module';
     AuthModule,
     FavoriteModule,
     ReportModule,
+    UserModule,
   ],
 })
 export class AppModule {}
