+++ src/auth/auth.module.ts
@@ -1,19 +1,17 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersRepository } from './repository/user.repository';
-import { RolesRepository } from './repository/role.repository';
 import { JwtModule } from '@nestjs/jwt';
 import { JwtStrategy } from './jwt.strategy';
 import { PassportModule } from '@nestjs/passport';
+import { UserModule } from 'src/user/user.module';
 
 @Module({
   controllers: [AuthController],
   imports: [
-    TypeOrmModule.forFeature([UsersRepository, RolesRepository]),
     JwtModule.register({ secret: '1234', signOptions: { expiresIn: 3600 } }),
     PassportModule.register({ defaultStrategy: 'jwt' }),
+    UserModule,
   ],
   providers: [AuthService, JwtStrategy],
   exports: [JwtStrategy, PassportModule],
