+++ src/auth/auth.service.ts
@@ -1,35 +1,26 @@
 import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
 import { SignUpDto } from './dto/signup.dto';
 import { SignInDto } from './dto/signin.dto';
-import { UsersRepository } from './repository/user.repository';
 import { UnauthorizedException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { JwtPayload } from './jwt-payload.interface';
 import * as bcrypt from 'bcrypt';
-import { RolesRepository } from './repository/role.repository';
-import { UserRole } from './role.enum';
-import { Role } from './entity/role.entity';
+import { UserService } from 'src/user/user.service';
 
 @Injectable()
 export class AuthService {
   constructor(
-    @InjectRepository(UsersRepository)
-    private usersRepository: UsersRepository,
-    @InjectRepository(RolesRepository) private rolesRepository: RolesRepository,
+    private userService: UserService,
     private jwtService: JwtService,
   ) {}
 
   async signUp(signUpDto: SignUpDto): Promise<void> {
-    const role: Role = await this.rolesRepository.findOne({
-      name: UserRole.USER,
-    });
-    return this.usersRepository.createUser(signUpDto, role);
+    return this.userService.createUser(signUpDto);
   }
 
   async signIn(signInDto: SignInDto): Promise<{ accessToken: string }> {
     const { username, password } = signInDto;
-    const user = await this.usersRepository.findOne({ username });
+    const user = await this.userService.findOne(username);
     if (user && (await bcrypt.compare(password, user.password))) {
       const payload: JwtPayload = { username };
       const accessToken: string = this.jwtService.sign(payload);
