+++ src/report/report.controller.ts
@@ -1,9 +1,9 @@
 import { Controller, Get, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { Roles } from 'src/auth/decorator/roles.decorator';
-import { User } from 'src/auth/entity/user.entity';
-import { UserRole } from 'src/auth/role.enum';
 import { Game } from 'src/game/game.entity';
+import { User } from 'src/user/entities/user.entity';
+import { UserRole } from 'src/user/role.enum';
 import { ReportService } from './report.service';
 
 @UseGuards(AuthGuard())
