+++ src/user/repositories/role.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Role } from './../entity/role.entity';
+import { Role } from '../entities/role.entity';
 
 @EntityRepository(Role)
 export class RolesRepository extends Repository<Role> {}
